Force XAudio2.0 -> 2.7 to use app-specific audio endpoints Application specific audio endpoints were only introduced in Windows 10 (1803) and they're not very well documented (that I can find, anyway), so there's an unfortunate number of assumptions in this implementation.
We COM hook XAudio2 from versions 2.0 -> 2.7 as these are what're used by games I've tested with.
These versions share the same interface (2.8+ don't, and apparently don't even use COM - but I didn't verify this), there's just some minor
differences with their constants which don't affect anything we care about here besides API defaults, but even then they don't matter for this patch.
The constants as-is are from 2.7, their final version before the interface changed.

IXAudio2::CreateMasteringVoice() takes in a default device index of 0.
There's no way of telling if we're explicitly stating we wish to use device index 0
or if we're wanting a default, but notably, internally (at kernel level somewhere? I couldn't observe the actual lookup) it will sometimes
(64-bit only perhaps, but even then it wasn't consistent) result in the correct default being used (application over system).

As such, when device index 0 is passed, we attempt to fetch the application-specific MM device ID (I'm not sure what this is actually called).
There doesn't seem to be any consistent version agnostic way of doing this that I could determine - even in 64 bit builds, when the correct
endpoint was ultimately used, I could not find a way of reporting which it was actually using internally.

Ultimately this means we resort to querying the registry directly for this information. Again, none of this is documented (that I could find),
so I was unable to determine what the application keys were (and even then they might not necessarily be used consistently), so we enumerate
all configured application audio endpoints instead.

Oddly, they're tied to DOS paths - which as far as I'm aware, aren't really used anymore, particularly over volume guid paths.
So, this means translating those paths back to regular NT paths, which, again, couldn't find a way to cleanly do, so we pull the drive name (e.g. D:)
and convert it to a DOS volume name (e.g. \Device\HarddiskVolume4), and just manually swap these. This may still break in other cases.

Matching the paths, we can identify the correct registry key, and then obtain what I refer to as an MM device path. This looks like:
\\?\SWD#MMDEVAPI#{0.0.0.00000000}.{fd9d7031-4c5d-4d34-be9b-0ec5a5168f8f}#{e6327cad-dcec-4949-ae8a-991e976a79d2}

In XAudio2.8, we'd be able to just this to specify the device, however in prior versions we need to translate this back to a numerical ID.
As it's not documented regarding just what this is (I can assume, but I'd rather not), to be safe we enumerate the available devices with XAudio2's
own API.
What it stores, however, is a portion of this - which is referred to as the "device ID". This looks like:
{0.0.0.00000000}.{fd9d7031-4c5d-4d34-be9b-0ec5a5168f8f}

Again, there's no documented clean way of handling this that I can find, so we unfortunately have to resort to string parsing to pull this.

Now we can match this against the device ID reported by XAudio2, and finally supply its associated device index to CreateMasteringVoice().

... All this because it doesn't consistently using application-specific endpoints for... some... reason...
And even when it does, it may choose not to on the next restart, for no discernible reason.

Yikes.